move everything to data layer, squash the migration
