+++ packages/frontpage/lib/data/db/vote.ts
@@ -2,7 +2,7 @@ import "server-only";
 import { getUser } from "../user";
 import { db } from "@/lib/db";
 import * as schema from "@/lib/schema";
-import { and, eq } from "drizzle-orm";
+import { and, eq, inArray, sql } from "drizzle-orm";
 import { cache } from "react";
 import { DID } from "../atproto/did";
 
@@ -43,11 +43,19 @@ export const getVoteForComment = cache(async (commentId: number) => {
 });
 
 export type UnauthedCreatePostVoteInput = {
-  postId: number;
   repo: DID;
+  rkey: string;
   hydratedVoteRecordValue: {
     createdAt: string;
-    cid: string;
+    subject: {
+      cid: string;
+      uri: {
+        rkey: string;
+        value: string;
+        authority: string;
+        collection: string;
+      };
+    };
   };
   hydratedRecord: {
     cid: string;
@@ -55,20 +63,36 @@ export type UnauthedCreatePostVoteInput = {
 };
 
 export const unauthed_createPostVote = async ({
-  postId,
   repo,
+  rkey,
   hydratedVoteRecordValue,
+  hydratedRecord,
 }: UnauthedCreatePostVoteInput) => {
   await db.transaction(async (tx) => {
+    const subject = (
+      await tx
+        .select()
+        .from(schema.Post)
+        .where(eq(schema.Post.rkey, hydratedVoteRecordValue.subject.uri.rkey))
+    )[0];
+
+    if (!subject) {
+      throw new Error(
+        `Subject not found with uri: ${hydratedVoteRecordValue.subject.uri.value}`,
+      );
+    }
+
+    if (subject.authorDid === repo) {
+      throw new Error(`[naughty] Cannot vote on own content ${repo}`);
+    }
     await tx.insert(schema.PostVote).values({
-      postId,
+      postId: subject.id,
       authorDid: repo,
       createdAt: new Date(hydratedVoteRecordValue.createdAt),
       cid: hydratedRecord.cid,
       rkey,
     });
 
-    console.log("Updating post aggregates");
     await tx
       .update(schema.PostAggregates)
       .set({
@@ -80,5 +104,151 @@ export const unauthed_createPostVote = async ({
       rank: sql<number>`
                 (CAST(COALESCE(${schema.PostAggregates.voteCount}, 1) AS REAL) / (pow((JULIANDAY('now') - JULIANDAY(${schema.PostAggregates.createdAt})) * 24 + 2,1.8)))`,
     });
+  });
+};
+
+export type UnauthedCreateCommentVoteInput = {
+  repo: DID;
+  rkey: string;
+  hydratedVoteRecordValue: {
+    createdAt: string;
+    subject: {
+      uri: {
+        authority: string;
+        collection: string;
+        rkey: string;
+        value: string;
+      };
+    };
+  };
+  hydratedRecord: {
+    cid: string;
+  };
+};
+
+export async function unauthed_createCommentVote({
+  repo,
+  rkey,
+  hydratedVoteRecordValue,
+  hydratedRecord,
+}: UnauthedCreateCommentVoteInput) {
+  await db.transaction(async (tx) => {
+    const subject = (
+      await tx
+        .select()
+        .from(schema.Comment)
+        .where(
+          eq(schema.Comment.rkey, hydratedVoteRecordValue.subject.uri.rkey),
+        )
+    )[0];
+
+    if (!subject) {
+      throw new Error(
+        `Subject not found with uri: ${hydratedVoteRecordValue.subject.uri.value}`,
+      );
+    }
+
+    if (subject.authorDid === repo) {
+      throw new Error(`[naughty] Cannot vote on own content ${repo}`);
+    }
+
+    await tx.insert(schema.CommentVote).values({
+      commentId: subject.id,
+      authorDid: repo,
+      createdAt: new Date(hydratedVoteRecordValue.createdAt),
+      cid: hydratedRecord.cid,
+      rkey,
+    });
+
+    const commentIds = tx
+      .select({ commentId: schema.Comment.id })
+      .from(schema.Comment)
+      .where(eq(schema.Comment.postId, subject.postId));
+
+    await tx
+      .update(schema.CommentAggregates)
+      .set({
+        voteCount: sql`${schema.CommentAggregates.voteCount} + 1`,
+      })
+      .where(eq(schema.CommentAggregates.commentId, subject.id));
+
+    await tx
+      .update(schema.CommentAggregates)
+      .set({
+        rank: sql<number>`
+                (CAST(COALESCE(${schema.CommentAggregates.voteCount}, 1) AS REAL) / (pow((JULIANDAY('now') - JULIANDAY(${schema.CommentAggregates.createdAt})) * 24 + 2,1.8)))`,
+      })
+      .where(inArray(schema.CommentAggregates.commentId, commentIds));
+  });
+}
+
+// Try deleting from both tables. In reality only one will have a record.
+// Relies on sqlite not throwing an error if the record doesn't exist.
+export const unauthed_deleteVote = async (rkey: string, repo: DID) => {
+  await db.transaction(async (tx) => {
+    const [commentTransaction] = await tx
+      .delete(schema.CommentVote)
+      .where(
+        and(
+          eq(schema.CommentVote.rkey, rkey),
+          eq(schema.CommentVote.authorDid, repo),
+        ),
+      )
+      .returning({
+        commentId: schema.CommentVote.commentId,
+      });
+
+    const [postVoteTransaction] = await tx
+      .delete(schema.PostVote)
+      .where(
+        and(
+          eq(schema.PostVote.rkey, rkey),
+          eq(schema.PostVote.authorDid, repo),
+        ),
+      )
+      .returning({ postId: schema.PostVote.postId });
+
+    if (commentTransaction?.commentId != null) {
+      //the vote is a comment vote
+
+      const postId = tx
+        .select({ postId: schema.Comment.postId })
+        .from(schema.Comment)
+        .where(eq(schema.Comment.id, commentTransaction.commentId));
+
+      const commentIds = tx
+        .select({ commentId: schema.Comment.id })
+        .from(schema.Comment)
+        .where(eq(schema.Comment.postId, postId));
+
+      await tx
+        .update(schema.CommentAggregates)
+        .set({
+          voteCount: sql`${schema.CommentAggregates.voteCount} - 1`,
+        })
+        .where(
+          eq(schema.CommentAggregates.commentId, commentTransaction.commentId),
+        );
+
+      await tx
+        .update(schema.CommentAggregates)
+        .set({
+          rank: sql<number>`
+                (CAST(COALESCE(${schema.CommentAggregates.voteCount}, 1) AS REAL) / (pow((JULIANDAY('now') - JULIANDAY(${schema.CommentAggregates.createdAt})) * 24 + 2,1.8)))`,
+        })
+        .where(inArray(schema.CommentAggregates.commentId, commentIds));
+    } else if (postVoteTransaction?.postId != null) {
+      //the vote is a post vote
+      await tx
+        .update(schema.PostAggregates)
+        .set({
+          voteCount: sql`${schema.PostAggregates.voteCount} - 1`,
+        })
+        .where(eq(schema.PostAggregates.postId, postVoteTransaction.postId));
+      await tx.update(schema.PostAggregates).set({
+        rank: sql<number>`
+                (CAST(COALESCE(${schema.PostAggregates.voteCount}, 1) AS REAL) / (pow((JULIANDAY('now') - JULIANDAY(${schema.PostAggregates.createdAt})) * 24 + 2,1.8)))`,
+      });
+    }
   });
 };
